gaialabsoftware/requirements.txt
appdirs==1.4.3
beautifulsoup4==4.5.3
Django==1.10.6
django-admin-tools==0.8.0
django-admin-tools-stats==0.8.0
django-bower==5.2.0
django-cache-utils==0.7.2
django-common-helpers==0.9.0
django-crontab==0.7.1
django-dbbackup==3.1.3
django-filter==1.0.2
django-jsonfield==1.0.1
django-nvd3==0.9.7
django-qsstats-magic==0.7.2
django-rest-auth==0.9.1
django-storages==1.5.2
djangorestframework==3.6.2
dropbox==7.2.1
Jinja2==2.9.6
lxml==3.7.3
MarkupSafe==1.0
olefile==0.44
packaging==16.8
Pillow==4.1.1
pyparsing==2.2.0
python-dateutil==2.6.0
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2017.2
reportlab==3.4.0
requests==2.13.0
six==1.10.0
typing==3.6.1
Unidecode==0.4.20
urllib3==1.20
